move train timeout check functions to utils
