+++ utils.py
@@ -2,7 +2,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from datetime import datetime
+from datetime import datetime, timedelta
 
 from smart_open import smart_open
 import requests
@@ -12,6 +12,17 @@ from requests.packages.urllib3.util.retry import Retry
 
 def utc_today():
     return datetime.utcnow().date()
+
+
+def retrain_timeout_expired(training_interval=timedelta(hours=24)):
+    with open('last_trained.txt', 'r') as f:
+        last_trained_time = datetime.strptime(f.read(), '%b %d %Y %H:%M')
+        return last_trained_time + training_interval < datetime.now()
+
+
+def save_training_time():
+    with open("last_trained.txt", "w") as f:
+        f.write(datetime.now().strftime('%b %d %Y %H:%M'))
 
 
 def get_with_retries(url, params=None, headers=None):
